Refactor: 분산된 북마크 Use Case 통합
