+++ src/main/java/com/vigilante/retriever/v1/bookmark/application/service/GetBookmarkService.java
@@ -6,18 +6,28 @@ import org.springframework.stereotype.Service;
 
 import com.vigilante.retriever.v1.bookmark.application.query.BookmarkMongoQuery;
 import com.vigilante.retriever.v1.bookmark.domain.entity.BookmarkEntity;
-import com.vigilante.retriever.v1.bookmark.domain.port.in.GetAllBookmarksUseCase;
+import com.vigilante.retriever.v1.bookmark.domain.port.in.GetBookmarkUseCase;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class GetAllBookmarksService implements GetAllBookmarksUseCase {
+public class GetBookmarkService implements GetBookmarkUseCase {
 
 	private final BookmarkMongoQuery bookmarkMongoQuery;
 
 	@Override
 	public List<BookmarkEntity> findAll() {
 		return bookmarkMongoQuery.findAll();
+	}
+
+	@Override
+	public BookmarkEntity getById(String bookmarkId) {
+		return bookmarkMongoQuery.getById(bookmarkId);
+	}
+
+	@Override
+	public List<BookmarkEntity> findByUserId(String userId) {
+		return bookmarkMongoQuery.findByUserId(userId);
 	}
 }
