--- src/main/java/com/vigilante/retriever/v1/bookmark/domain/port/in/GetBookmarkByIdUseCase.java
@@ -1,7 +1,0 @@
-package com.vigilante.retriever.v1.bookmark.domain.port.in;
-
-import com.vigilante.retriever.v1.bookmark.domain.entity.BookmarkEntity;
-
-public interface GetBookmarkByIdUseCase {
-	BookmarkEntity getById(String bookmarkId);
-}
