+++ src/main/java/com/vigilante/retriever/v1/bookmark/domain/port/in/GetBookmarkUseCase.java
@@ -4,6 +4,11 @@ import java.util.List;
 
 import com.vigilante.retriever.v1.bookmark.domain.entity.BookmarkEntity;
 
-public interface GetBookmarksByUserUseCase {
+public interface GetBookmarkUseCase {
+
+	List<BookmarkEntity> findAll();
+
+	BookmarkEntity getById(String bookmarkId);
+
 	List<BookmarkEntity> findByUserId(String userId);
 }
